pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvisbase"
version="0.0.1"
description = "JarvisBase Library"
dependencies = [
    "opencv-python==4.7.0.72",
    "setuptools>=49.2.0",
    "tqdm>=4.32.2",
    "numpy==1.21.6",
    "requests>=2.20.0",
    "ipython>=7.5.0",
    "typing>=3.6.6",
    "lxml>=4.3.3",
    "psutil>=5.6.2",
    "Pyro4>=4.76",
    "getch>=1.0; sys_platform != 'win32' and sys_platform != 'cygwin'",
    "coloredlogs>=10.0",
    "pillow>=8.0.0",
    "dill>=0.3.1.1",
    "daemoniker>=0.2.3",
    "xmltodict==0.12.0",
    "inflection>=0.3.1",
    "jinja2>=2.11.2",
    "imagehash>=4.0.0",
    "av",
    "clip", 
    "attr==0.3.2",
    "attrs==22.1.0",
    "dm_tree==0.1.7",
    "gym3==0.3.3",
    "gymnasium",
    "hydra_colorlog", 
    "hydra-core==1.3.1",
    "hydra-joblib-launcher",
    "lmdb==1.3.0",
    "matplotlib==3.5.3",
    "pandas==1.4.3",
    "Pillow==9.4.0",
    "seaborn==0.12.2",
    "transformers==4.21.1",
    "tree==0.2.4",
    "wandb==0.13.3",
    "tensorboardX",
    "deepspeed",
    "tabulate",
    "cchardet",
    "torchvision",
    "torchmetrics",
    "efficientnet_pytorch", 
    "pytorch-lightning==2.0.6",
    "scipy==1.8.0",
    "flaky",
    "pyglet",
    "pynput",
    "redis", 
    "rich",
    "ipdb",
    "ray[default]==2.4.0",
    "watermark",
    "tensorflow_probability",
    "lz4",
    "tensorboard",
    "mlflow",
    "kornia",
    "aligo",
    "chardet", 
]
authors = [
    {name="CraftJarvis", email="craftjarvis@outlook.com"},
]
[options]
package_data = {'jarvis'= ['*.json']}

[options.packages.find]
where = "jarvis"

[tool.setuptools.dynamic]
readme = {file = "README.md"}


[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
